improve endpoints

+++ src/index.js
@@ -38,7 +38,7 @@ const authenticateJWT = (req, res, next) => {
     if (!token) {
       return res.status(401).json({
         success: false,
-        message: 'Token de autenticación no proporcionado',
+        message: 'Authentication token not provided',
       });
     }
 
@@ -46,7 +46,7 @@ const authenticateJWT = (req, res, next) => {
       if (err) {
         return res.status(403).json({
           success: false,
-          message: 'Token de autenticación inválido',
+          message: 'Invalid authentication token',
         });
       }
 
@@ -63,7 +63,7 @@ app.use(authenticateJWT);
 
 const port = 4500;
 app.listen(port, () => {
-  console.log(`Servidor iniciado en http://localhost:${port}`);
+  console.log(`Server started at http://localhost:${port}`);
 });
 
 //1.ENDPOINT PARA COGER TODOS LOS HECHIZOS
@@ -88,41 +88,54 @@ app.get('/spells', async (req, res) => {
 app.post('/spells', async (req, res) => {
   const dataSpell = req.body; //objeto
   const {
-    slug,
+    name,
     category,
-    creator,
     effect,
-    hand,
+    hand_movement,
     image,
     incantation,
     light,
-    name,
     wiki,
   } = dataSpell;
 
+  if (
+    !name ||
+    !category ||
+    !effect ||
+    !light ||
+    name.trim() === '' ||
+    category.trim() === '' ||
+    effect.trim() === '' ||
+    light.trim() === ''
+  ) {
+    return res.status(400).json({
+      success: false,
+      message:
+        'Upps! Spells not added. Name, category, effect, and light are required fields.',
+    });
+  }
+
   let sql =
-    'INSERT INTO spells (slug, category, creator, effect, hand, image,incantation, light, name,wiki) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?);';
+    'INSERT INTO spells (name, category, effect, hand_movement, image,incantation, light, wiki) VALUES (?, ?, ?, ?, ?, ?, ?, ?);';
 
   try {
     const conn = await getConnection();
 
     const [spellsList] = await conn.query(sql, [
-      slug,
+      name,
       category,
-      creator,
       effect,
-      hand,
+      hand_movement,
       image,
       incantation,
       light,
-      name,
       wiki,
     ]);
 
     if (spellsList.affectedRows === 0) {
       res.json({
         success: false,
-        message: 'No se ha podido insertar',
+        message: 'It was not possible to add the spell',
       });
       return;
     }
@@ -132,10 +145,21 @@ app.post('/spells', async (req, res) => {
       id: spellsList.insertId,
     });
   } catch (error) {
-    res.json({
-      success: false,
-      message: `Ha ocurrido un error${error}`,
-    });
+    // Captura el error de restricción única
+    if (error.code === 'ER_DUP_ENTRY') {
+      res.json({
+        success: false,
+        message:
+          'It was not possible to add the spell. The name is already in use.',
+      });
+    } else {
+      res.json({
+        success: false,
+        message: `An error has occurred while adding the spell: ${error}`,
+      });
+    }
+  } finally {
+    conn.end();
   }
 });
 
@@ -144,42 +168,83 @@ app.post('/spells', async (req, res) => {
 app.put('/spells/:id', async (req, res) => {
   const dataSpell = req.body; //objeto
   const {
-    slug,
+    name,
     category,
-    creator,
     effect,
-    hand,
+    hand_movement,
     image,
     incantation,
     light,
-    name,
     wiki,
   } = dataSpell;
 
   const idSpell = req.params.id;
 
   let sql =
-    'UPDATE spells SET slug=?, category=?, creator=?, effect=?, hand=?, image=?, incantation=?, light=?, name=?, wiki=? WHERE idSpell=?';
+    'UPDATE spells SET name=?, category=?, effect=?, hand_movement=?, image=?, incantation=?, light=?, wiki=? WHERE idSpell=?;';
 
   const conn = await getConnection();
 
-  const [spellsList] = await conn.query(sql, [
-    slug,
-    category,
-    creator,
-    effect,
-    hand,
-    image,
-    incantation,
-    light,
-    name,
-    wiki,
-    idSpell,
-  ]);
+  try {
+    const [spellsList] = await conn.query(sql, [
+      name,
+      category,
+      effect,
+      hand_movement,
+      image,
+      incantation,
+      light,
+      wiki,
+      idSpell,
+    ]);
 
+    res.json({
+      success: true,
+      message: 'Spell updated successfully',
+    });
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({
+      success: false,
+      message: 'An error has occurred while updating the spell.',
+    });
+  } finally {
+    conn.end();
+  }
+});
+
+//EXTRA, MOSTRAR UN HECHIZO
+
+app.get('/spells/:id', async (req, res) => {
+  const idSpell = req.params.id;
+
+  if (isNaN(parseInt(idSpell))) {
+    res.json({
+      success: false,
+      error: 'To find your spell the id must be a number',
+    });
+    return;
+  }
+
+  let query = 'SELECT * FROM spells WHERE idSpell =?';
+
+  const conn = await getConnection();
+
+  //Ejecutar esa consulta
+  const [spellsList] = await conn.query(query, [idSpell]);
+  const numOfSpells = spellsList.length;
+
+  if (numOfSpells === 0) {
+    res.json({
+      success: true,
+      message: 'The spell you are looking for does not exist.',
+    });
+    return;
+  }
+
+  //Enviar una respuesta
   res.json({
-    success: true,
-    message: 'Actualizado correctamente',
+    spellsList: spellsList[0], // listado
   });
 });
 
@@ -188,16 +253,35 @@ app.put('/spells/:id', async (req, res) => {
 app.delete('/spells/:id', async (req, res) => {
   const idSpell = req.params.id;
 
+  if (!idSpell || isNaN(idSpell)) {
+    return res
+      .status(400)
+      .json({ success: false, message: 'Invalid spell ID.' });
+  }
+
   let sql = 'DELETE FROM spells WHERE idSpell=?';
 
   const conn = await getConnection();
 
-  const [spellList] = await conn.query(sql, [idSpell]);
+  try {
+    const [spellList] = await conn.query(sql, [idSpell]);
 
-  res.json({
-    success: true,
-    message: 'Eliminado correctamente',
-  });
+    if (spellList.affectedRows === 0) {
+      return res
+        .status(404)
+        .json({ success: false, message: 'Spell not found.' });
+    }
+
+    res.json({
+      success: true,
+      message: 'Spell deleted successfully',
+    });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ success: false, message: 'Internal Server Error' });
+  } finally {
+    conn.end();
+  }
 });
 
 //5. BONUS: REGISTRO DE USUARIOS
@@ -207,9 +291,7 @@ app.post('/register', async (req, res) => {
 
   try {
     if (!email || !password || !name) {
-      throw new Error(
-        'Se requiere un correo electrónico, una contraseña y un nombre.'
-      );
+      throw new Error('To register an email, password, and name are required.');
     }
     const passwordHashed = await bcrypt.hash(password, 10);
 
@@ -226,9 +308,10 @@ app.post('/register', async (req, res) => {
       token: token,
     });
   } catch (error) {
-    res.json({
+    console.error(error);
+    return res.status(500).json({
       success: false,
-      error: error.message,
+      error: 'An error has occurred while registering the user.',
     });
   }
 });
@@ -241,31 +324,42 @@ app.post('/login', async (req, res) => {
   let sql = 'SELECT * FROM users WHERE email =?';
 
   const conn = await getConnection();
-  const [user] = await conn.query(sql, [email]);
 
-  if (user.length === 0) {
-    return res.status(401).json({
-      success: false,
-      message: 'Credenciales incorrectas',
+  try {
+    const [user] = await conn.query(sql, [email]);
+
+    if (user.length === 0) {
+      return res.status(401).json({
+        success: false,
+        message: 'Incorrect credentials',
+      });
+    }
+
+    const isValidPassword = await bcrypt.compare(password, user[0].password);
+
+    if (!isValidPassword) {
+      return res.status(401).json({
+        success: false,
+        message: 'Incorrect login credentials',
+      });
+    }
+
+    const tokenPayload = { idUser: user[0].idUser, email: user[0].email };
+    const token = generateToken(tokenPayload);
+
+    res.json({
+      success: true,
+      token: token,
     });
-  }
-
-  const isValidPassword = await bcrypt.compare(password, user[0].password);
-
-  if (!isValidPassword) {
-    return res.status(401).json({
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({
       success: false,
-      message: 'Credenciales incorrectas',
+      message: 'An error has occurred while attempting to log in.',
     });
+  } finally {
+    conn.end();
   }
-
-  const tokenPayload = { idUser: user[0].idUser, email: user[0].email };
-  const token = generateToken(tokenPayload);
-
-  res.json({
-    success: true,
-    token: token,
-  });
 });
 
 //6.BONUS. MIDDLEWARE DE AUTENTICACIÓN (definido en L.29)
@@ -276,15 +370,26 @@ app.get('/profile', authenticateJWT, async (req, res) => {
   const email = req.user.email;
 
   const conn = await getConnection();
-  const [userData] = await conn.query('SELECT * FROM users WHERE idUser=?', [
-    idUser,
-  ]);
 
-  res.json({
-    success: true,
-    profile: {
-      email: email,
-      additionalData: userData,
-    },
-  });
+  try {
+    const [userData] = await conn.query('SELECT * FROM users WHERE idUser=?', [
+      idUser,
+    ]);
+
+    res.json({
+      success: true,
+      profile: {
+        email: email,
+        additionalData: userData,
+      },
+    });
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({
+      success: false,
+      message: 'An error has occurred while fetching user profile data.',
+    });
+  } finally {
+    conn.end();
+  }
 });
